Implementations

+++ src/main/java/com/web/dssapp/repository/UserRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.mongodb.repository.Query;
 
 import com.web.dssapp.model.User;
 
-public interface UserRepository extends MongoRepository<User, Long> {
+public interface UserRepository extends MongoRepository<User, Integer> {
 	
 	@Query("{name:'?0'}")
     User findUserByName(String name);
